% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR.R
\name{odataR_get_data}
\alias{odataR_get_data}
\title{Get (part of) an OData data structure}
\usage{
odataR_get_data(root = odataR_get_root(), table_id = NULL, query = NULL,
  save_file_name = NULL)
}
\arguments{
\item{root}{Root of data structure}

\item{table_id}{Identification of table}

\item{query}{OData query to restrict data returned from structure}

\item{save_file_name}{Name of file to save the XML data in or NULL}
}
\description{
Determines from which structure the data will be extracted
}
\examples{
\dontrun{
odataR_get_data(table_id='82935NED')
odataR_get_data(root='http://opendata.cbs.nl/ODataFeed/OData',table_id='82935NED')
}
}

