% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{odataR_get_root}
\alias{odataR_get_root}
\title{Gets the root for OData data structure}
\usage{
odataR_get_root()
}
\description{
Determines from which structure the data will be extracted
}
\examples{
odataR_get_root()

}

