% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_subtable}
\alias{odataR_get_subtable}
\title{Convert a sub table (contents or properties) to a data.frame}
\usage{
odataR_get_subtable(dsn, mt = "table", query = NULL, save_XML = NULL)
}
\arguments{
\item{dsn}{Named character vector with full name of subtable}

\item{mt}{Indicates destination of the XML data. Character \code{'table'} (default) or \code{'prop'} or \code{NULL}}

\item{query}{OData query to restrict data returned from structure, See
\code{\link{odataR_get_table}} for examples of queries}

\item{save_XML}{Name of file to save the XML data in or NULL or default name}
}
\value{
A data.frame with the contents of the sub table (\code{mt='table'}) or the properties of the table (\code{mt='prop'}) or the XML contents of the sub table (\code{mt=NULL})
}
\description{
Convert a  sub table (contents or properties) to a data.frame. The function calls \code{\link{odataR_get_data}} one or more times (the latter if the server returns more than 10000 rows) to actually retrieve the XML data. With  parameter \code{mt} one indicates that the requests relates to a 'normal' sub table or relates to the properties sub table. In these cases the data is converted to a data.frame . By specifying \code{mt=NULL} the contents is not converted to a data.frame but returned as XML.
}
\examples{
\dontrun{
subtabs = odataR_get_subtables(table_id="82935NED")
}
}

