% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR.R
\name{odataR_get_subtables}
\alias{odataR_get_subtables}
\title{Get information about sub tables in OData data structure}
\usage{
odataR_get_subtables(root = odataR_get_root(), table_id = NULL)
}
\arguments{
\item{root}{Root of data structure}

\item{table_id}{Identification of table}
}
\description{
Get information about sub tables in OData data structure
}
\examples{
\dontrun{
odataR_get_subtables(table_id='82935NED')
odataR_get_subtables(root='http://opendata.cbs.nl/ODataFeed/OData',table_id='82935NED')
}
}

