% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_list_tables}
\alias{odataR_list_tables}
\title{List the tables in an OData structure}
\usage{
odataR_list_tables(root = odataR_get_root_catalog(), query = NULL)
}
\arguments{
\item{root}{Root of data structure}

\item{query}{OData query to restrict data returned from structure, See
\code{\link{odataR_get_table}} for examples of queries}
}
\value{
A data.frame when successful otherwise \code{NULL}
}
\description{
List the tables in an OData structure
}
\examples{
\dontrun{
odataR_set_root("http://dataderden.cbs.nl")
df_cat = odataR_list_tables()
RIVM   = odataR_list_tables(query="?$filter=Catalog eq 'RIVM'")
}
}

