% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_data}
\alias{odataR_get_data}
\title{Get (part of) an OData data structure}
\usage{
odataR_get_data(root = odataR_get_root(), table_id = NULL, query = NULL,
  save_file_name = NULL)
}
\arguments{
\item{root}{Root of data structure}

\item{table_id}{Identification of table}

\item{query}{OData query to restrict data returned from structure, See
\code{\link{odataR_get_table}} for examples of queries}

\item{save_file_name}{Name of file to save the XML data in or NULL}
}
\description{
This is the only function in the package that actually acquires data from the database. This function returns a block of at most 10000 rows (restriction of the interface) in XML format. It is therefore also suitable in cases where the result is not a nice table (e.g. in the case of TypeInfo). The function \code{\link{odataR_get_table}} will call this function one or more times (depending on the size of the data) and convert the combined data to a table or property structure.
}
\examples{
\dontrun{
odataR_get_data(table_id='82935NED')
odataR_get_data(root='http://opendata.cbs.nl/ODataFeed/OData',table_id='82935NED')
}
}

