% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR.R
\name{odataR_get_subtable}
\alias{odataR_get_subtable}
\title{Copy table (or properties) to a data.frame}
\usage{
odataR_get_subtable(dsn, mt = "table", query = NULL, save_XML = NULL)
}
\arguments{
\item{dsn}{Named character vector with full name of subtable}

\item{mt}{Character 'table' or 'prop' or NULL}

\item{query}{OData query to restrict data returned from structure}

\item{save_XML}{Name of file to save the XML data in or NULL or default name}
}
\description{
Copy table to a data.frame
}
\examples{
\dontrun{
subtabs = odataR_get_subtables(table_id="82935NED")
}
}

