% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataR_code.R
\name{odataR_get_table}
\alias{odataR_get_table}
\title{Get table in decoded form}
\usage{
odataR_get_table(root = odataR_get_root(), table_id = NULL, query = NULL,
  typed = T, keepcode = c())
}
\arguments{
\item{root}{Root of data structure}

\item{table_id}{Identification of table}

\item{query}{OData query to restrict data returned from structure}

\item{typed}{Boolean indicating 'TypedDataSet' when T or 'UntypedDataSet' when F'}

\item{keepcode}{Character string with dimension(s) for which the coded values are kept}
}
\description{
The main function is \code{\link{odataR_get_table}} that extracts the information from the database and decodes the dimensions. It uses the other functions in the package behind the screens.
E.g. when executing \code{df = odataR_get_table(table_id="82935NED")} the following is executed:
\itemize{
 \item get names of sub tables: \cr
 \code{subtabs = odataR_get_subtables(table_id="82935NED")}
 \item retrieve coded data: \cr
 \code{df      = odataR_get_subtable(subtabs['TypedDataSet'])}
 \item retrieve properties: \cr
 \code{props   = odataR_get_subtable(subtabs['DataProperties'],mt='prop')}
 \item decode dimensions: \cr
 \code{df      = couple_data(df, ...)}
}
}
\section{Remark}{

See \url{http://docs.oasis-open.org/odata/odata/v4.0/errata02/os/complete/part1-protocol/odata-v4.0-errata02-os-part1-protocol-complete.html} for details about the query possibilities
}
\examples{
\dontrun{
df      = odataR_get_table(table_id="82935NED")
df      = odataR_get_table(table_id="82935NED",keepcode = "RegioS")
df      = odataR_get_table(table_id="82935NED",keepcode = c("RegioS","Perioden"))
df      = odataR_get_table(table_id="82935NED",keepcode = "RegioS",
      query  = "?$format=atom&$filter=startswith(RegioS,'NL01')" )
df      = odataR_get_table(table_id="82935NED",
  query  = paste0("?$format=atom&$filter=startswith(RegioS,'NL01')",
                  "&$select=RegioS,Perioden,TotaleInvesteringen_1") )
df      = odataR_get_table(table_id="82935NED",
  query  = paste0("?$format=atom&$filter=startswith(RegioS,'NL01')",
                  "&$select=RegioS,Perioden,TotaleInvesteringen_1",
                  "&$skip=2&$top=3") )
}
}

