% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{odataR_set_root}
\alias{odataR_set_root}
\title{Sets the root for OData data structure}
\usage{
odataR_set_root(root = NULL)
}
\arguments{
\item{root}{NULL for the default or the url otherwise}
}
\description{
Determines from which structure the data will be extracted
}
\examples{
odataR_set_root()
odataR_set_root('http://opendata.cbs.nl/ODataFeed/OData')
}

